Binary Index Tree: Implementation of Binary Index Tree in Python3
Binary Index Tree enables us to get the sum of elemnts of array of a range
in log(nn) time instead of n. We are creating a Binary Index Tree which contains
sum of their child elements and the parent node is  find by adding the last set bit
of the iterator to the iterator.  While getting the sum from the tree the parent node or
parent elment is found by substracting the last set bit of iterator to the iterator.

Signed-off-by: satyam kumar<satyamkumar4@acm.org>